⚡️ Label components 만들어서 모듈화 해주기
label components에서 눈 모양 아이콘 js 가져다가 쓰도록 하기
