+++ login.html
@@ -40,6 +40,6 @@
     </div>
 
   </main>
-  <script src="./js/login.js"></script>
+  <script type="module" src="./js/login.js"></script>
 </body>
 </html>
