+++ signup.html
@@ -21,12 +21,12 @@
       <label for="signup-password">비밀번호</label>
       <div class="signup-password-container">
         <input type="password" name="signup-password" id="signup-password" class="input" placeholder="비밀번호를 입력해주세요">
-        <img alt="btn_visibility" src="./img/btn_visibility_off.svg" class="signup-password-visibility"/>
+        <img alt="btn_visibility" src="img/btn_visibility_off.svg" class="signup-password-visibility"/>
       </div>
       <label for="signup-check-password">비밀번호 확인</label>
       <div class="signup-checkpassword-container">
         <input type="password" name="signup-check-password" id="signup-check-password" class="input" placeholder="비밀번호를 다시 한 번 입력해주세요">
-        <img alt="btn_visibility" src="./img/btn_visibility_off.svg" class="signup-checkpassword-visibility"/>
+        <img alt="btn_visibility" src="img/btn_visibility_off.svg" class="signup-checkpassword-visibility"/>
       </div>
       <button class="signup-btn">
         회원가입
@@ -47,6 +47,6 @@
     </div>
 
   </main>
-  <script src="./js/signup.js"></script>
+  <script type="module" src="./js/signup.js"></script>
 </body>
 </html>
